A poul 2 weeks

+++ 0x00-python_variable_annotations/2-floor.py
@@ -3,11 +3,11 @@ import math
 '''find floor of a number'''
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     '''
     floor any king of n value
         Args:
         n (number): a entere float number
     Return the floor of a number
     '''
-    return math.floor(n)
+    return int(math.floor(n))
